==> 5. Detection and correction validation.ipynb <==
import pandas as pd
import numpy as np
import re
from transformers import BertTokenizer, BertForMaskedLM
import torch
from transformers import AdamW
from tqdm import tqdm
import gensim
from gensim.models import Word2Vec 
from gensim.models import KeyedVectors
import pandas as pd
from nltk.tokenize import RegexpTokenizer
import pickle
import ast
import statistics as s
from fuzzywuzzy import fuzz
from nltk.corpus import stopwords
import nltk
from transformers import pipeline
import copy
rom nltk.corpus import stopwords

# load word2vec model
#w2v_model = ...
# load BERT model
#BERT_model = ...
# load dataframe
#df = ...

tokenizer = BertTokenizer.from_pretrained("GroNLP/bert-base-dutch-cased")
BERT_model = BertForMaskedLM.from_pretrained("GroNLP/bert-base-dutch-cased")
#w2v_model.intersect_word2vec_format(r"combined-160.txt", binary=False, lockf=1.0)
# https://github.com/clips/dutchembeddings

df = pd.read_csv('preprocessed_df100.csv')

df.columns

# create vocabulary lists
vocab_BERT = list(tokenizer.vocab.keys())
print(len(vocab_BERT))
train = df[df['set']=='train']
#df['gt for training'] = df['gt text']
train = '.'.join(list(train['gt for training']))
train = (train.replace('.', '')).split(' ')
vocab_BERT = vocab_BERT + train
vocab_word2vec = list(model.vocab.keys())
print(len(vocab_word2vec))
vocab_word2vec = vocab_word2vec + train
print(len(vocab_BERT))
print(len(vocab_word2vec))

# create list of historical expressions

def finetune_word2vec(train, window=5):
    sentences = train.split('.')
    tokenizer = RegexpTokenizer(r'\w+')
    sentences = [tokenizer.tokenize(i) for i in sentences]
    total_examples = len(sentences)
    
    model_w2v = Word2Vec(size=160, min_count=1, window=window)
    model_w2v.build_vocab(sentences)
    total_examples = model_w2v.corpus_count
    model = KeyedVectors.load_word2vec_format(r"combined-160.txt", binary=False)
    model_w2v.build_vocab([list(model.vocab.keys())], update=True)
    model_w2v.intersect_word2vec_format(r"combined-160.txt", binary=False, lockf=1.0)
    model_w2v.train(sentences, total_examples=total_examples, epochs=model_w2v.iter)
    return model_w2v

train = df['gt text'][0]
word2vec_model = finetune_word2vec(train)

df.columns
print(df.index.values.tolist())

# set hyperparameters for validation
#topn_detection = 1000
#topn_correction = 1000
#window = 5

def list_merger(lists):
    normal_list = False
    for elem in lists:
        if type(elem) != list:
            normal_list = True
    if normal_list == True:
        return lists
    else:
        new_list = []
        for elem in lists:
            new_list = new_list + elem
        return new_list

#lijst = [['aa','bb','cc'],['dd','ee','ff'],[7,8,9]]
lijst = [['engelant', 'ampc'], ['dublin', 'den', '24', 'maert'], ['tegens', 'de', 'charters', 'van', 'dese', 'stadt'], []]
lijst = list_merger(lijst)
print(lijst)

string = "[['engelant', 'ampc'], ['dublin', 'den', '24', 'maert'], ['tegens', 'de', 'charters', 'van', 'dese', 'stadt'], []]"
lijst = ast.literal_eval(string)
print(type(lijst))
print(lijst)
print(lijst[0][0])

# skiplist (words that should not be corrected: names)
with open("ocr_names.txt", "rb") as fp:   # Unpickling
    ocr_names = pickle.load(fp)

ocr_names = []
for name in ocr_names:
    if len(name) >= 5:
        ocr_names.append(name)


def correct_sorted(candidates, sim_or_probs, LD): # sorts first by LD, then by similarity/probability
    paired_sorted = sorted(zip(LD,sim_or_probs,candidates),key = lambda x: (x[0],x[1]), reverse=True)
    LD,sim_or_probs,candidates = zip(*paired_sorted)
    correction = candidates[0]
    return correction
    
def correct_calculated(candidates, sim_or_probs, LD): # calculates a score from LD and normalised similarity/probability
    inv_LD = 1 - LD
    sim_or_probs = np.array(sim_or_probs)
    sim_or_probs = np.interp(sim_or_probs, (sim_or_probs.min(), sim_or_probs.max()), (0, 1)).tolist()
    score = sim_or_probs / inv_LD
    zipped_pairs = zip(score.tolist(), candidates)
    sorted_by_score = [x for _, x in sorted(zipped_pairs, reverse=True)]
    correction = sorted_by_score[0]
    return correction

def remove_stopwords(candidates, cosine, LD):
    #nltk.download('stopwords')
    stop_words = set(stopwords.words('dutch'))
    candidates_nostopwords = []
    cosine_nostopwords = []
    LD_nostopwords = []
    for i in range(len(candidates)):
        if candidates[i] not in stop_words:
            candidates_nostopwords.append(candidates[i])
            cosine_nostopwords.append(cosine[i])
            LD_nostopwords.append(LD[i])
    LD_nostopwords = np.array(LD_nostopwords)
    return candidates_nostopwords, cosine_nostopwords, LD_nostopwords

#detection and correction validation word2vec
error_positions = [] # the position of a word when it is an error in predictions
non_error_positions = [] # the position of a word when it is not an error in predictions
right_token_positions = []

rights_correct_sorted_list = []
wrongs_correct_sorted_list = []
rights_correct_sorted_nosw_list = []
wrongs_correct_sorted_nosw_list = []
rights_correct_calculated_list = []
wrongs_correct_calculated_list = []

def detection_and_correction_word2vec(row, w2v_model, ocr_names, window=5, topn_detection=1000):
    if row['set'] != 'val':
        return np.nan
    else:
        print('start')
        identifier = row['identifier']
        OCR_text = row['aligned_OCR_sentences']
        GT_text = row['aligned_GT_sentences']
        OCR_text = ast.literal_eval(OCR_text)
        GT_text = ast.literal_eval(GT_text)
        OCR_text = list_merger(OCR_text)
        GT_text = list_merger(GT_text)
        window_range = list(range(0,window))
        window_range = np.array(window_range) - ((window - 1) / 2)
        
        # keep track of positions in candidates
        error_positions_doc = []
        non_error_positions_doc = []
        right_token_positions_doc = []
        
        # keep track of performance
        rights_correct_sorted = 0
        wrongs_correct_sorted = 0
        rights_correct_sorted_nosw = 0
        wrongs_correct_sorted_nosw = 0
        rights_correct_calculated = 0
        wrongs_correct_calculated = 0
        
        for i in range(len(OCR_text)):
            try:
                if (OCR_text[i] in ocr_names) or (OCR_text[i].isalpha() == False) or (len(OCR_text[i]) <= 2):
                    continue
                error = True
                if OCR_text[i] == GT_text[i]:
                    error = False
                context = []
                for j in window_range:
                    if (i+j >= 0) and (i+j < len(OCR_text)) and j != 0:
                        context.append(OCR_text[i+int(j)])
                    else:
                        pass
                candidates = []
                cosines = []
                # calculate positions detection task
                for prediction in w2v_model.predict_output_word(context, topn=topn_detection):
                    candidates.append(prediction[0])
                    cosines.append(prediction[1]) 
                # remove punctuation except for hyphen from candidates
                candidates = [re.sub(r'[^\w\d\s\-]+', '', x) for x in candidates]
                try:
                    position = candidates.index(OCR_text[i])
                except ValueError:
                    position = topn_detection
                if error == True:
                    # where the error is in the detection list
                    error_positions_doc.append(position)
                    # find where the right word is in the candidates list (same as detection list)
                    try: 
                        position_right_token = candidates.index(GT_text[i])
                    except ValueError: 
                        position_right_token = topn_detection
                    print('position_right_token:', position_right_token)
                    right_token_positions_doc.append(position_right_token)
                    # try two correction methods
                    # first calculate the normalized LDs:
                    LD = np.array([fuzz.ratio(OCR_text[i], word)/100 for word in candidates])
                    # try sorting method
                    correction = correct_sorted(candidates, cosines, LD)
                    if correction == GT_text[i]:
                        rights_correct_sorted += 1
                    elif correction != GT_text[i]:
                        wrongs_correct_sorted += 1
                    print('OCR:', OCR_text[i])
                    print('correction sorted:', correction)
                    # try again the sorting methods, but without stopwords
                    candidates_nostopwords, cosine_nostopwords, LD_nostopwords = remove_stopwords(candidates, cosines, LD)
                    correction = correct_sorted(candidates_nostopwords, cosine_nostopwords, LD_nostopwords)
                    if correction == GT_text[i]:
                        rights_correct_sorted_nosw += 1
                    elif correction != GT_text[i]:
                        wrongs_correct_sorted_nosw += 1
                    # try score calculation method
                    correction = correct_calculated(candidates, cosines, LD)
                    if correction == GT_text[i]:
                        rights_correct_calculated += 1
                    elif correction != GT_text[i]:
                        wrongs_correct_calculated += 1
                    print('OCR:', OCR_text[i])
                    print('correction score:', correction)
                elif error == False:
                    # where the non error is in the detection list
                    non_error_positions_doc.append(position)
                # calculate positions correction task
                if error == True:
                    try:
                        right_token_position = candidates.index(GT_text[i])
                        right_token_positions_doc.append(right_token_position)
                    except ValueError:
                        right_token_positions_doc.append(topn_detection)
                    right_token_positions.append(right_token_positions_doc)
            except:
                pass
            
            
    #return error_positions, non_error_positions
    error_positions.append(error_positions_doc)
    non_error_positions.append(non_error_positions_doc)
    right_token_positions.append(right_token_positions_doc)
    
    # add performance to the list
    rights_correct_sorted_list.append(rights_correct_sorted)
    wrongs_correct_sorted_list.append(wrongs_correct_sorted)
    rights_correct_sorted_nosw_list.append(rights_correct_sorted_nosw)
    wrongs_correct_sorted_nosw_list.append(wrongs_correct_sorted_nosw)
    rights_correct_calculated_list.append(rights_correct_calculated)
    wrongs_correct_calculated_list.append(wrongs_correct_calculated)
    
    return identifier, error_positions, non_error_positions, right_token_positions_doc, rights_correct_sorted_list, wrongs_correct_sorted_list, rights_correct_sorted_nosw_list, wrongs_correct_sorted_nosw_list, rights_correct_calculated_list, wrongs_correct_calculated_list   
    
#for index, row in df.iterrows():
#    detection_word2vec(row)
# df.loc[70]
fake_test_list_GT_aligned = """Een koekenpan of kortweg pan is een platte pan met een lang handvat.
De pan ontleent zijn naam aan het feit dat in zo'n pan pannenkoeken worden gebakken. Ook ander voedsel, zoals vlees, wordt in een koekenpan gebraden"""
fake_test_list_OCR_aligned = """Een hoekenpan of kortweg pan is een platte pan met een hang handvat.
De pan ontleent zijn naam haan het feit dat in zo'n pan pannenkoeken horden gebakken. Ook ander voedsel, zoals vlees, word in een hoekenpan gebraden"""
fake_test_list_GT_aligned = fake_test_list_GT_aligned.split('.')
fake_test_list_OCR_aligned = fake_test_list_OCR_aligned.split('.')
fake_test_list_GT_aligned = [x.split(' ') for x in fake_test_list_GT_aligned]
fake_test_list_OCR_aligned = [x.split(' ') for x in fake_test_list_OCR_aligned]
d = {'identifier': ['111'], 'aligned_OCR_sentences': [str(fake_test_list_OCR_aligned)], 'aligned_GT_sentences': [str(fake_test_list_GT_aligned)], 'set': ['val']}
df_probeer = pd.DataFrame(data=d)

for index, row in df.loc[df['set'].isin(['val'])].iterrows():
    if (index == 70) or (index == 72):
        print('index:', index)
        identifier, error_positions, non_error_positions, right_token_positions_doc, rights_correct_sorted_list, wrongs_correct_sorted_list, rights_correct_sorted_nosw_list, wrongs_correct_sorted_nosw_list, rights_correct_calculated_list, wrongs_correct_calculated_list = detection_and_correction_word2vec(row, word2vec_model, ocr_names)




print(rights_correct_sorted_list)
print(wrongs_correct_sorted_list)
print(rights_correct_sorted_nosw_list)
print(wrongs_correct_sorted_nosw_list)
print(rights_correct_calculated_list)
print(wrongs_correct_calculated_list)

# analysis of performance on validation set word2vec
#d = {'error_positions in detection': error_positions, 'non_error_positions in detection': non_error_positions}
#validation_df_positions = pd.DataFrame(data = d)
d = {'rights_correct_sorted_list': rights_correct_sorted_list, 'wrongs_correct_sorted_list': wrongs_correct_sorted_list, \
    'rights_correct_sorted_nosw_list': rights_correct_sorted_nosw_list, 'wrongs_correct_sorted_nosw_list': wrongs_correct_sorted_nosw_list, \
    'rights_correct_calculated_list': rights_correct_calculated_list, 'wrongs_correct_calculated_list': wrongs_correct_calculated_list}
validation_df_correction_methods_word2vec = pd.DataFrame(data = d)


print(error_positions)
print(non_error_positions)
#print(error_positions)
error_positions = list_merger(error_positions)
non_error_positions = list_merger(non_error_positions)
right_token_positions = list_merger(right_token_positions)
print('errors:', s.mean(error_positions))
print('non errors:', s.mean(non_error_positions))
print('right token in prediction:', s.mean(right_token_positions))
print(validation_df_correction_methods_word2vec.head())

df.columns

#detection and correction validation BERTje
error_positions = [] # the position of a word when it is an error in predictions
non_error_positions = [] # the position of a word when it is not an error in predictions
right_token_positions = []

rights_correct_sorted_list = []
wrongs_correct_sorted_list = []
rights_correct_sorted_nosw_list = []
wrongs_correct_sorted_nosw_list = []
rights_correct_calculated_list = []
wrongs_correct_calculated_list = []

def detection_and_correction_BERTje(row, BERT_model, tokenizer, ocr_names, topn_detection=1000):
    if row['set'] != 'val':
        return np.nan
    else:
        identifier = row['identifier']
        print('Val')
        OCR_text = row['aligned_OCR_sentences']
        GT_text = row['aligned_GT_sentences']
        print('OCR:', OCR_text)
        print('GT:', GT_text)
        OCR_text = ast.literal_eval(OCR_text)
        GT_text = ast.literal_eval(GT_text)
        print(OCR_text)
        
        # keep track of positions in candidates
        error_positions_doc = []
        non_error_positions_doc = []
        right_token_positions_doc = []
        
        # keep track of performance
        rights_correct_sorted = 0
        wrongs_correct_sorted = 0
        rights_correct_sorted_nosw = 0
        wrongs_correct_sorted_nosw = 0
        rights_correct_calculated = 0
        wrongs_correct_calculated = 0
        
        print(OCR_text)
        
        for j in range(len(OCR_text)):
            for i in range(len(OCR_text[j])):
                if (OCR_text[j][i] in ocr_names) or (OCR_text[j][i].isalpha() == False) or (len(OCR_text[j][i]) <= 2):
                    continue
                error = True
                if OCR_text[j][i] == GT_text[j][i]:
                    error = False
                candidates = []
                probabilities = []
                # calculate positions detection task
                sentence = copy.deepcopy(OCR_text[j])
                sentence[i] = '[MASK]'
                sentence = ' '.join(sentence)
                print(sentence)
                pipe = pipeline('fill-mask', model=BERT_model, tokenizer = tokenizer, top_k=topn_detection)
                for res in pipe(sentence):
                    candidates.append(res['token_str'].replace(' ', ''))
                    probabilities.append(res['score'])
                # remove punctuation except for hyphen from candidates
                candidates = [re.sub(r'[^\w\d\s\-]+', '', x) for x in candidates]
                try:
                    position = candidates.index(OCR_text[j][i])
                except ValueError:
                    position = topn_detection
                if error == True:
                    # where the error is in the detection list
                    error_positions_doc.append(position)
                    # find where the right word is in the candidates list (same as detection list)
                    try: 
                        position_right_token = candidates.index(GT_text[j][i])
                    except ValueError: 
                        position_right_token = topn_detection
                    print('position_right_token:', position_right_token)
                    right_token_positions_doc.append(position_right_token)
                    # try two correction methods
                    # first calculate the normalized LDs:
                    LD = np.array([fuzz.ratio(OCR_text[j][i], word)/100 for word in candidates])
                    # try sorting method
                    correction = correct_sorted(candidates, probabilities, LD)
                    if correction == GT_text[j][i]:
                        rights_correct_sorted += 1
                    elif correction != GT_text[j][i]:
                        wrongs_correct_sorted += 1
                    print('OCR:', OCR_text[j][i])
                    print('correction sorted:', correction)
                    # try again the sorting methods, but without stopwords
                    candidates_nostopwords, probabilities_nostopwords, LD_nostopwords = remove_stopwords(candidates, probabilities, LD)
                    correction = correct_sorted(candidates_nostopwords, probabilities_nostopwords, LD_nostopwords)
                    if correction == GT_text[j][i]:
                        rights_correct_sorted_nosw += 1
                    elif correction != GT_text[j][i]:
                        wrongs_correct_sorted_nosw += 1
                    # try score calculation method
                    correction = correct_calculated(candidates, probabilities, LD)
                    if correction == GT_text[j][i]:
                        rights_correct_calculated += 1
                    elif correction != GT_text[j][i]:
                        wrongs_correct_calculated += 1
                    print('OCR:', OCR_text[j][i])
                    print('correction score:', correction)
                elif error == False:
                    # where the non error is in the detection list
                    non_error_positions_doc.append(position)
                # calculate positions correction task
                if error == True:
                    try:
                        right_token_position = candidates.index(GT_text[j][i])
                        right_token_positions_doc.append(right_token_position)
                    except ValueError:
                        right_token_positions_doc.append(topn_detection)
                    right_token_positions.append(right_token_positions_doc)
            
            
    #return error_positions, non_error_positions
    error_positions.append(error_positions_doc)
    non_error_positions.append(non_error_positions_doc)
    right_token_positions.append(right_token_positions_doc)
    
    # add performance to the list
    rights_correct_sorted_list.append(rights_correct_sorted)
    wrongs_correct_sorted_list.append(wrongs_correct_sorted)
    rights_correct_sorted_nosw_list.append(rights_correct_sorted_nosw)
    wrongs_correct_sorted_nosw_list.append(wrongs_correct_sorted_nosw)
    rights_correct_calculated_list.append(rights_correct_calculated)
    wrongs_correct_calculated_list.append(wrongs_correct_calculated)
    
    return identifier, error_positions, non_error_positions, right_token_positions_doc, rights_correct_sorted_list, wrongs_correct_sorted_list, rights_correct_sorted_nosw_list, wrongs_correct_sorted_nosw_list, rights_correct_calculated_list, wrongs_correct_calculated_list   
    
#for index, row in df.iterrows():
#    detection_word2vec(row)
# df.loc[70]
fake_test_list_GT_aligned = """Een koekenpan of kortweg pan is een platte pan met een lang handvat.
De pan ontleent zijn naam aan het feit dat in zo'n pan pannenkoeken worden gebakken. Ook ander voedsel, zoals vlees, wordt in een koekenpan gebraden"""
fake_test_list_OCR_aligned = """Een hoekenpan of kortweg pan is een platte pan met een hang handvat.
De pan ontleent zijn naam haan het feit dat in zo'n pan pannenkoeken horden gebakken. Ook ander voedsel, zoals vlees, word in een hoekenpan gebraden"""
fake_test_list_GT_aligned = fake_test_list_GT_aligned.split('.')
fake_test_list_OCR_aligned = fake_test_list_OCR_aligned.split('.')
fake_test_list_GT_aligned = [x.split(' ') for x in fake_test_list_GT_aligned]
fake_test_list_OCR_aligned = [x.split(' ') for x in fake_test_list_OCR_aligned]
d = {'identifier': ['111'], 'aligned_OCR_sentences': [str(fake_test_list_OCR_aligned)], 'aligned_GT_sentences': [str(fake_test_list_GT_aligned)], 'set': ['val']}
df_probeer = pd.DataFrame(data=d)
identifier, error_positions, non_error_positions, right_token_positions_doc, rights_correct_sorted_list, wrongs_correct_sorted_list, rights_correct_sorted_nosw_list, wrongs_correct_sorted_nosw_list, rights_correct_calculated_list, wrongs_correct_calculated_list = detection_and_correction_BERTje(df_probeer.loc[0], BERT_model, tokenizer, ocr_names)




# analysis of performance on validation set BERTje
#d = {'error_positions in detection': error_positions, 'non_error_positions in detection': non_error_positions}
#validation_df_positions = pd.DataFrame(data = d)
d = {'rights_correct_sorted_list': rights_correct_sorted_list, 'wrongs_correct_sorted_list': wrongs_correct_sorted_list, \
    'rights_correct_sorted_nosw_list': rights_correct_sorted_nosw_list, 'wrongs_correct_sorted_nosw_list': wrongs_correct_sorted_nosw_list, \
    'rights_correct_calculated_list': rights_correct_calculated_list, 'wrongs_correct_calculated_list': wrongs_correct_calculated_list}
validation_df_correction_methods_BERTje = pd.DataFrame(data = d)

print(right_token_positions)
#print(error_positions)
error_positions = list_merger(error_positions)
non_error_positions = list_merger(non_error_positions)
right_token_positions = list_merger(right_token_positions)
print('errors:', s.mean(error_positions))
print('non errors:', s.mean(non_error_positions))
print('right token in prediction:', s.mean(right_token_positions))
validation_df_correction_methods_BERTje.head()

text = "Dit1 is2 een3 zin4 die5 lang6 genoeg7 is8"
text = text.split(' ')

window = 5
window_range = list(range(0,window))
print(window_range)
window_range = np.array(window_range) - ((window - 1) / 2)
print(window_range)

context = []

for i in range(len(text)):
    print(i)
    print(text[i])
    context = []
    for j in window_range:
        if (i+j >= 0) and (i+j < len(text)) and j != 0:
            context.append(text[i+int(j)])
        else:
            pass
    print('context:', context)
    



word2vec_model.predict_output_word(['Een', 'of', 'kortweg'], topn=10)

word2vec_model.wv.most_similar('papier')

pipe = pipeline('fill-mask', model=BERT_model, tokenizer = tokenizer, top_k=10)
for res in pipe('Ik vind jou een [MASK].'):
    print(res['sequence'])
    print(res)

old_sentence = "Die [MASK] hadde viele avontuere ghemaect."
new_sentence = "De [MASK] had vele avonturen gemaakt."
pipe = pipeline('fill-mask', model=BERT_model, tokenizer = tokenizer, top_k=5)
for res in pipe(new_sentence):
    print(res['token_str'].replace(' ', ''))
    print(res['score'])
    pipe = pipeline('fill-mask', model=BERT_model, tokenizer = tokenizer, top_k=5)
for res in pipe(old_sentence):
    print(res['token_str'].replace(' ', ''))
    print(res['score'])
